requirements.txt
streamlit
onnxruntime
torch
torchvision
numpy
pillow
opencv-python-headless>=4.5.0
scipy
matplotlib
grad-cam
plotly
sentence-transformers
faiss-cpu
duckduckgo_search